+++ src/isntRaw.c
@@ -3,3 +3,7 @@
 bool isntRaw(SEXP x) {
   return TYPEOF(x) != RAWSXP;
 }
+
+bool isRaw(SEXP x) {
+  return TYPEOF(x) == RAWSXP;
+}